requirements.txt
kagglehub
numpy
flask
Pillow
torch==2.6.0
torchvision==0.21.0
transformers
seaborn
pandas
# pip install -e . --index-url https://download.pytorch.org/whl/cpu